requirements.txt
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
notion-client==2.2.1
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
urllib3==2.2.1
Werkzeug==3.0.2
